SAK-13584
   Patch from Chuck
"
This will build on SAK-11544 and SAK-12837. The idea is to add an even better feature to the Message API and then modify the
MailArchive tool to use this new API. The improvements will support flexible filtering, sorting, and paging all in a single
query object. This will be done in a very similar manner to the GenericDAO search pattern that is currently in use by Aaron Z.
"


git-svn-id: https://source.sakaiproject.org/svn/kernel@49932 66ffb92e-73f9-0310-93c1-f5514f145a0a