KNL-12  Patch applied from Joe Dane
From orrigianal SAK
If a user uploads a series of resources with names like "slide1.jpg", "slide2.jpg", ..., "slide10.jpg", ..., the natural lexicographic sort will be sub-optimal, sorting slide10.jpg between slide1.jpg and slide2.jpg. This can be worked-around (by using names like "slide01.jpg", etc.) but this is typically not intuitive.

The patch changes the comparator used by the CHS to notice these "mixed-mode" names, and to sort appropriately. In the patch, the new sort behavior is enabled by default but can be switched-off.


git-svn-id: https://source.sakaiproject.org/svn/kernel@51225 66ffb92e-73f9-0310-93c1-f5514f145a0a