KNL-15 SAK-12458
Applied patch from Aaron
One of the Spring masters suggested this to me today. In order to
reduce database load, you put a lazy proxy in between the datasource
and the users (consumers) of it. This means the actual connection is only opened
when it is used and if there are things (like hibernate or some
transactional wrappers) which always open a connection then it does
not waste the connection if the hibernate cache has the value or the
connection is not actually used. This has been around since Spring 1.1 is
highly tested and reliable.

Here is the link to the email exchange from the sakai dev list:
http://www.nabble.com/Datasource-proxy-for-efficiency-td14305747.html


git-svn-id: https://source.sakaiproject.org/svn/kernel@51237 66ffb92e-73f9-0310-93c1-f5514f145a0a