KNL-21 SAK-14219
Patch provided by 	Marty Banting
======================================
This is a result of a context.redirect() method call by the tool with a relative URL. The context.redirect() invocation ultimately results in a HttpServletResponseWrapper.sendRedirect() method call. According to the Javadocs, sendRedirect

"Sends a temporary redirect response to the client using the specified redirect location URL. This method can accept relative URLs; the servlet container must convert the relative URL to an absolute URL before sending the response to the client. If the location is relative without a leading '/' the container interprets it as relative to the current request URI. If the location is relative with a leading '/' the container interprets it as relative to the servlet container root. "

From analysis, it seems that from ActiveToolComponent.MyActiveTool.WrappedResponse.sendRedirect(), the URIs used are:
- Tomcat: m_res.response.response.request.coyoteRequest.decodedUriMB.strValue
- Websphere: m_res.m_req.request._srtRequestHelper._requestURI

Unfortunately, the URI value is different between Tomcat and MySQL. It would be something similar to:
- Tomcat: /portal/tool/590b5975-613d-4c73-bc17-82c77ff339f6/compose
- Websphere: /sakai-messageforums-tool/jsp/compose.jsf

For Websphere, this results in "/sakai-messageforums-tool/jsp/sakai.filepicker.helper/tool", which is an invalid URL in Sakai.

The fix is to modify ActiveToolComponent.MyActiveTool.WrappedResponse.sendRedirect() to convert the given target relative URL (eg. "sakai.filepicker.helper/tool") to an absolute URL using the current context path and path info, if deployed on Websphere. Tomcat behavior is not affected and stays the same.

Fix was implemented and tested in branch https://source.sakaiproject.org/svn//kernel/branches/SAK-13408/ and the corresponding patch is attached here.
Tested and confirmed fix works on both Tomcat and Websphere.
=========================================


git-svn-id: https://source.sakaiproject.org/svn/kernel@51431 66ffb92e-73f9-0310-93c1-f5514f145a0a