KNL-75
Patch from Cris Holdorph

As documented in this area of Sakai confluence

http://confluence.sakaiproject.org/confluence/display/TERRA/Home

This is the work that enables Sakai sessions to be clustered using Terracotta. Additional work must still be done for any tool that would like to support clustering. This is also documented in the confluence space above.

Thanks

git-svn-id: https://source.sakaiproject.org/svn/kernel@54470 66ffb92e-73f9-0310-93c1-f5514f145a0a