KNL-67
Patch from  	 Savitha

Admin tools like Realms, Sites, Users and Alias takes unusual time to load the page. The reason is because, the existing query, it scan over the entire data set (in oracle database), though it displays only 20 records at a time.
---
The problem has been resolved and a patch is attached as kernel-util-patch.txt.

These changes are part of kernel. The sql query is changed for better efficiency, which scans only a part of the data it displays, and it takes less than a second to display the page, as against ~10s before.

Thanks

git-svn-id: https://source.sakaiproject.org/svn/kernel@54476 66ffb92e-73f9-0310-93c1-f5514f145a0a