KNL-221 Commit change from Matthew Buckett.
It's unsafe to check for values set in sakai.properties in a filter init() block, because when the first webapp loads, the rest of Sakai will not have started, so this could fail for the first webapp and succeed for the others.

Because the load order of webapps is not known, if the first webapp is portal or something equally important, this could lead to session handling failure.

git-svn-id: https://source.sakaiproject.org/svn/kernel@64283 66ffb92e-73f9-0310-93c1-f5514f145a0a