KNL-1091 Allow services to be notified of site removal.
This adds an advisor interface that allows services to know when a site is going to be removed and take some action.
The site is still around which means that if some information about the site is needed it can still be found. Listening for site removal events means that the service only finds a site has been removed once it is gone.

git-svn-id: https://source.sakaiproject.org/svn/kernel@306930 66ffb92e-73f9-0310-93c1-f5514f145a0a