KNL-1310 Remove old copy of deleted resource when deleting.
When the recycle bin is enabled and there is already a copy of a the file in the recycle bin we now delete it. This is because there is no way with the ContentHosting API to specify which file should be restored and it appears to vary between DBs.

In the future we may wish to revert this change and add an API method that allows you to restore a file by UUID which would allow users to pick which version of a deleted file they wish to restore.

git-svn-id: https://source.sakaiproject.org/svn/kernel@315736 66ffb92e-73f9-0310-93c1-f5514f145a0a