KNL-1194 If we fail to un-delete a file always unlock it.
Previously if a file was being restored and it failed (eg file exists or would push site over quota) then the resource in the recycle bin wouldn't get unlocked and so could never be restored. Now we always make sure to unlock it at the end of a restore.

Thanks to Juan José Meroño Sánchez for the patch.

git-svn-id: https://source.sakaiproject.org/svn/kernel@315738 66ffb92e-73f9-0310-93c1-f5514f145a0a